AB3649 - Moved manifest stuff
