+++ src/Umbraco.Abstractions/Manifest/IManifestFilter.cs
