+++ src/Umbraco.Abstractions/Manifest/ManifestFilterCollection.cs
