+++ src/Umbraco.Abstractions/Manifest/ManifestFilterCollectionBuilder.cs
